+++ eventstore/mongodb/spring/reactor/src/test/java/org/occurrent/eventstore/mongodb/spring/reactor/ReactorMongoEventStoreTest.java
@@ -33,6 +33,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledForJreRange;
 import org.junit.jupiter.api.condition.EnabledOnJre;
+import org.junit.jupiter.api.condition.EnabledOnOs;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.occurrent.cloudevents.OccurrentCloudEventExtension;
 import org.occurrent.condition.Condition;
@@ -71,6 +72,7 @@ import java.util.stream.Stream;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.time.ZoneOffset.UTC;
 import static java.util.Objects.requireNonNull;
+import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
 import static org.hamcrest.Matchers.not;
@@ -78,6 +80,7 @@ import static org.hamcrest.Matchers.nullValue;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.condition.JRE.JAVA_11;
 import static org.junit.jupiter.api.condition.JRE.JAVA_8;
+import static org.junit.jupiter.api.condition.OS.MAC;
 import static org.occurrent.cloudevents.OccurrentCloudEventExtension.STREAM_VERSION;
 import static org.occurrent.condition.Condition.*;
 import static org.occurrent.eventstore.api.SortBy.SortDirection.ASCENDING;
@@ -729,7 +732,8 @@ public class ReactorMongoEventStoreTest {
         @DisplayName("parallel writes")
         class ParallelWritesToEventStoreReturns {
 
-            @RepeatedIfExceptionsTest(repeats = 10, suspend = 500)
+            @EnabledOnOs(MAC)
+            @RepeatedIfExceptionsTest(repeats = 5, suspend = 500)
             void parallel_writes_to_event_store_throws_WriteConditionNotFulfilledException() {
                 // Given
                 CyclicBarrier cyclicBarrier = new CyclicBarrier(2);
@@ -750,7 +754,7 @@ public class ReactorMongoEventStoreTest {
                 }).start();
 
                 // Then
-                Awaitility.await().untilAsserted(() -> assertThat(exception.get()).isEqualTo(new WriteConditionNotFulfilledException("name", 0, writeCondition, "ikk")));
+                Awaitility.await().atMost(4, SECONDS).untilAsserted(() -> assertThat(exception).hasValue(new WriteConditionNotFulfilledException("name", 1, writeCondition, "WriteCondition was not fulfilled. Expected version to be equal to 0 but was 1.")));
             }
         }